% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pdf}
\alias{get_pdf}
\title{All PDF data}
\usage{
get_pdf(path_x, choose.file = FALSE)
}
\arguments{
\item{path_x}{write the path of your file.}

\item{choose.file}{choose.file = FALSE means you want all data files, choose.file = TRUE means you want specific data files.}
}
\value{
A list
}
\description{
get pdf files and list it.
}
\seealso{
Check \code{\href{https://github.com/BYUIDSS/Rbills}{README.md}} to know how to use.
}
