% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pdf_want}
\alias{get_pdf_want}
\title{Choose data you want}
\usage{
get_pdf_want(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A list
}
\description{
Choose A List Of Files Interactively On MS Windows
}
\examples{
getwd()
get_pdf_want(getwd())
get_pdf_want("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/power_bills")
}
