% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbills_functions.r
\name{read_all_powerbills}
\alias{read_all_powerbills}
\title{Scrape all power bill PDF's in a folfer/directory}
\usage{
read_all_powerbills(x)
}
\arguments{
\item{x}{x is a path to a direcotry}
}
\value{
A CSV or Table
}
\description{
Takes in a working direcotry which contains PDF files and scrapes them
}
