% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_gas}
\alias{read_pdf_gas}
\title{Read All Gas PDF Files}
\usage{
read_pdf_gas(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A table
}
\description{
Read All Gas PDF Files
}
\examples{
c <- "C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/Gas_bills_plain_text/"
read_pdf_gas("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/Gas_bills_plain_text")
}
