% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_gas_want}
\alias{read_pdf_gas_want}
\title{Read Gas PDF Files You Want}
\usage{
read_pdf_gas_want(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A table
}
\description{
Read Gas PDF Files You Want
}
\examples{
c <- "C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/Gas_bills_plain_text/"
read_pdf_gas_want("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/Gas_bills_plain_text")
}
