% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_seg}
\alias{read_pdf_seg}
\title{Read All Gas PDF Files}
\usage{
read_pdf_seg(path_x = getwd(), x)
}
\arguments{
\item{path_x}{write the path of your file.}

\item{x}{put the file's list}
}
\value{
A table
}
\description{
Read All Gas PDF Files
}
\examples{
devtools::load_all()
pdf_file <- system.file("data-raw", "example_gasbill.pdf", package = "Rbills", mustWork = TRUE)
pdf_folder <- system.file("data-raw", package = "Rbills", mustWork = TRUE)
read_pdf_seg(pdf_folder, x)
read_pdf_seg("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_package/Rbills/data-raw", x)
}
