% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_seg_want}
\alias{read_pdf_seg_want}
\title{Read Gas PDF Files You Want}
\usage{
read_pdf_seg_want(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A table
}
\description{
Read Gas PDF Files You Want
}
\examples{
pdf_file <- system.file("data-raw", "example_gasbill.pdf", package = "Rbills", mustWork = TRUE)
pdf_folder <- system.file("data-raw", package = "Rbills", mustWork = TRUE)
read_pdf_seg_want(pdf_folder)
}
