% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pdf}
\alias{get_pdf}
\title{All PDF data}
\usage{
get_pdf(path_x, choose.file = FALSE)
}
\arguments{
\item{path_x}{write the path of your file.}

\item{choose.file}{choose.file = FALSE means you want all data files, choose.file = TRUE means you want specific data files.}
}
\value{
A list
}
\description{
get pdf files and list it.
}
\examples{
devtools::load_all()
pdf_folder <- system.file("data-raw", package = "Rbills", mustWork = TRUE)
get_pdf(pdf_folder)
get_pdf(pdf_folder, choose.file = TRUE)
}
