% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pdf_all}
\alias{get_pdf_all}
\title{All PDF data}
\usage{
get_pdf_all(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A list
}
\description{
get pdf files and list it.
}
\examples{
get_pdf_all("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/power_bills")

}
