% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_rmp}
\alias{read_pdf_rmp}
\title{Read All Power PDF Files}
\usage{
read_pdf_rmp(path_x = getwd(), x)
}
\arguments{
\item{path_x}{write the path of your file.}

\item{x}{put the file's list}
}
\value{
A table
}
\description{
Read All Power PDF Files
}
\seealso{
Check \code{\href{https://github.com/BYUIDSS/Rbills}{README.md}} to know how to use.
}
