% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_rmp}
\alias{read_pdf_rmp}
\title{Read All Power PDF Files}
\usage{
read_pdf_rmp(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A table
}
\description{
Read All Power PDF Files
}
\examples{
c <- "C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/power_bills/"
read_pdf_rmp("C:/Users/User/Desktop/MATH 488 Brother Hathaway/merit_medical_FA19/documents/reference_material/power_bills")
}
