% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_rmp}
\alias{read_pdf_rmp}
\title{Read All Power PDF Files}
\usage{
read_pdf_rmp(path_x, x, building_names = NULL)
}
\arguments{
\item{path_x}{a string that sets the directory where the filenames are located.}

\item{x}{a vector of strings that have filenames.}

\item{building_names}{defaults to NULL.  If NULL then user is prompted to type names.  Else a vector of building names can be provided to avoid the prompt.}
}
\value{
A table
}
\description{
Read All Power PDF Files
}
\seealso{
Check \href{https://github.com/BYUIDSS/Rbills}{README.md} to know how to use.
}
