% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_rmp}
\alias{read_pdf_rmp}
\title{Read All Power PDF Files}
\usage{
read_pdf_rmp(x)
}
\arguments{
\item{x}{write the path of your file.}
}
\value{
A table
}
\description{
Read All Power PDF Files
}
\examples{
devtools::load_all()
pdf_file <- system.file("data-raw", "example_powerbill.pdf", package = "Rbills", mustWork = TRUE)
pdf_folder <- system.file("data-raw", package = "Rbills", mustWork = TRUE)
read_pdf_rmp(pdf_folder)
}
