% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pdf_seg}
\alias{read_pdf_seg}
\title{Read All Gas PDF Files}
\usage{
read_pdf_seg(path_x, x)
}
\arguments{
\item{path_x}{a string that sets the directory where the filenames are located.}

\item{x}{a vector of strings that have filenames.}
}
\value{
A table
}
\description{
Read All Gas PDF Files
}
\seealso{
Check \href{https://github.com/BYUIDSS/Rbills}{README.md} to know how to use.
}
